backend/requirements.txt
fastapi
uvicorn
jinja2
pydantic
weasyprint
markdown
openai
jsonschema
cryptography
requests
python-dotenv